Reorganized the dependency graph to follow Dagger's guidelines on testing
